django/data/requirements.txt
Django>=4.2
djangorestframework
pydicom
requests
djangorestframework-simplejwt
psycopg2-binary
django-filter